requirements.txt
django
docx
docxtpl
django-import_export
pillow